+++ src/backend/utils/activity/pgstat_replslot.c
@@ -16,7 +16,7 @@
  * dropped while shut down, which is addressed by not restoring stats for
  * slots that cannot be found by name when starting up.
  *
- * Copyright (c) 2001-2024, PostgreSQL Global Development Group
+ * Copyright (c) 2001-2025, PostgreSQL Global Development Group
  *
  * IDENTIFICATION
  *	  src/backend/utils/activity/pgstat_replslot.c
@@ -194,8 +194,8 @@ pgstat_replslot_to_serialized_name_cb(const PgStat_HashKey *key, const PgStatSha
 	 * at the offset.
 	 */
 	if (!ReplicationSlotName(key->objid, name))
-		elog(ERROR, "could not find name for replication slot index %llu",
-			 (unsigned long long) key->objid);
+		elog(ERROR, "could not find name for replication slot index %" PRIu64,
+			 key->objid);
 }
 
 bool
