+++ src/oracle_test/modules/test_slru/test_slru.c
@@ -8,7 +8,7 @@
  * Portions Copyright (c) 2023-2025, IvorySQL Global Development Team
  *
  * IDENTIFICATION
- *		src/test/modules/test_slru/test_slru.c
+ *		src/oracle_test/modules/test_slru/test_slru.c
  *
  *	 add the file for requirement "SQL PARSER"
  * -------------------------------------------------------------------------
@@ -153,8 +153,8 @@ test_slru_page_sync(PG_FUNCTION_ARGS)
 	ftag.segno = pageno / SLRU_PAGES_PER_SEGMENT;
 	SlruSyncFileTag(TestSlruCtl, &ftag, path);
 
-	elog(NOTICE, "Called SlruSyncFileTag() for segment %lld on path %s",
-		 (long long) ftag.segno, path);
+	elog(NOTICE, "Called SlruSyncFileTag() for segment %" PRId64 " on path %s",
+		 ftag.segno, path);
 
 	PG_RETURN_VOID();
 }
@@ -168,8 +168,8 @@ test_slru_page_delete(PG_FUNCTION_ARGS)
 	ftag.segno = pageno / SLRU_PAGES_PER_SEGMENT;
 	SlruDeleteSegment(TestSlruCtl, ftag.segno);
 
-	elog(NOTICE, "Called SlruDeleteSegment() for segment %lld",
-		 (long long) ftag.segno);
+	elog(NOTICE, "Called SlruDeleteSegment() for segment %" PRId64,
+		 ftag.segno);
 
 	PG_RETURN_VOID();
 }
